[SPE-1026] Introduce EvervaultEnclaves package
