+++ Sources/EvervaultCages/EvervaultAttestation/PcrManager.swift
@@ -6,6 +6,8 @@
 //
 
 import Foundation
+import EvervaultCore
+import EvervaultCages
 
 public struct StoredPcrProvider {
     public var pcrs: [PCRs]
@@ -23,7 +25,7 @@ class PcrManager {
     private let backgroundQueue = DispatchQueue(label: "com.evervault.pcrManagerQueue")
     
     private init() {
-        print("Evervault: Started polling Cage PCR providers every \(refreshInterval) seconds")
+        print("Evervault: Started polling PCR providers every \(refreshInterval) seconds")
         scheduleManagerUpdate()
     }
     
@@ -53,60 +55,60 @@ class PcrManager {
     }
     
     private func fetchPcrsFromAllStoredProviders() {
-        for cageIdentifier in self.store.keys {
-            runProvider(cageIdentifier: cageIdentifier) { _, error in
+        for identifier in self.store.keys {
+            runProvider(identifier: identifier) { _, error in
                 if let error = error {
-                    print("Evervault: Failed to update PCRs for \(cageIdentifier): \(error.localizedDescription)")
+                    print("Evervault: Failed to update PCRs for \(identifier): \(error.localizedDescription)")
                 } 
             }
         }
     }
     
-    private func runProvider(cageIdentifier: String, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
-        guard let storedProvider = self.store[cageIdentifier] else {
-            completion(nil, NSError(domain: "Evervault", code: 404, userInfo: [NSLocalizedDescriptionKey: "No provider found for \(cageIdentifier)."]))
+    private func runProvider(identifier: String, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
+        guard let storedProvider = self.store[identifier] else {
+            completion(nil, NSError(domain: "Evervault", code: 404, userInfo: [NSLocalizedDescriptionKey: "No provider found for \(identifier)."]))
             return
         }
 
-        attemptProviderWithRetries(cageIdentifier: cageIdentifier, storedProvider: storedProvider, retries: 3, delay: 1, completion: completion)
+        attemptProviderWithRetries(identifier: identifier, storedProvider: storedProvider, retries: 3, delay: 1, completion: completion)
     }
 
-    private func attemptProviderWithRetries(cageIdentifier: String, storedProvider: StoredPcrProvider, retries: Int, delay: TimeInterval, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
+    private func attemptProviderWithRetries(identifier: String, storedProvider: StoredPcrProvider, retries: Int, delay: TimeInterval, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
         
         storedProvider.provider { [weak self] newPCRs, error in
             guard let self = self else { return }
 
             if let newPCRs = newPCRs {
                 let updatedProvider = StoredPcrProvider(pcrs: newPCRs, provider: storedProvider.provider)
-                self.store[cageIdentifier] = updatedProvider
-                print("Evervault: Successfully updated PCRs for \(cageIdentifier) from provider.")
+                self.store[identifier] = updatedProvider
+                print("Evervault: Successfully updated PCRs for \(identifier) from provider.")
                 completion(updatedProvider, nil)
             } else if retries > 0 {
                 // If there's an error and we still have retries left, try again after the specified delay.
-                print("Evervault: Error fetching PCRs for \(cageIdentifier): \(String(describing: error)). Retrying in \(delay) seconds...")
+                print("Evervault: Error fetching PCRs for \(identifier): \(String(describing: error)). Retrying in \(delay) seconds...")
                 DispatchQueue.global().asyncAfter(deadline: .now() + delay) {
-                    self.attemptProviderWithRetries(cageIdentifier: cageIdentifier, storedProvider: storedProvider, retries: retries - 1, delay: delay * 2, completion: completion)
+                    self.attemptProviderWithRetries(identifier: identifier, storedProvider: storedProvider, retries: retries - 1, delay: delay * 2, completion: completion)
                 }
             } else {
-                print("Evervault: Exhausted retries for \(cageIdentifier). Unable to fetch PCRs.")
-                completion(nil, error ?? NSError(domain: "Evervault", code: 500, userInfo: [NSLocalizedDescriptionKey: "Exhausted retries for \(cageIdentifier)."]))
+                print("Evervault: Exhausted retries for \(identifier). Unable to fetch PCRs.")
+                completion(nil, error ?? NSError(domain: "Evervault", code: 500, userInfo: [NSLocalizedDescriptionKey: "Exhausted retries for \(identifier)."]))
             }
         }
     }
     
-    private func registerPcrProviderAndFetch(cageIdentifier: String, provider: @escaping (@escaping ([PCRs]?, Error?) -> Void) -> Void, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
+    private func registerPcrProviderAndFetch(identifier: String, provider: @escaping (@escaping ([PCRs]?, Error?) -> Void) -> Void, completion: @escaping (StoredPcrProvider?, Error?) -> Void) {
             let storedProvider = StoredPcrProvider(pcrs: [], provider: provider)
-            self.store[cageIdentifier] = storedProvider
-            runProvider(cageIdentifier: cageIdentifier, completion: completion)
+            self.store[identifier] = storedProvider
+            runProvider(identifier: identifier, completion: completion)
     }
 
-    func fetchPcrs(cageIdentifier: String, provider: @escaping (@escaping ([PCRs]?, Error?) -> Void) -> Void, completion: @escaping ([PCRs]?, Error?) -> Void) {
-        if let storedPcrProvider = self.store[cageIdentifier] {
+    func fetchPcrs(identifier: String, provider: @escaping (@escaping ([PCRs]?, Error?) -> Void) -> Void, completion: @escaping ([PCRs]?, Error?) -> Void) {
+        if let storedPcrProvider = self.store[identifier] {
             // If the provider is already stored, use it to fetch PCRs
             storedPcrProvider.provider(completion)
         } else {
             // If the provider isn't stored yet, register and fetch
-            registerPcrProviderAndFetch(cageIdentifier: cageIdentifier, provider: provider) { storedProvider, error in
+            registerPcrProviderAndFetch(identifier: identifier, provider: provider) { storedProvider, error in
                 if let storedProvider = storedProvider {
                     completion(storedProvider.pcrs, nil)
                 } else if let error = error {
